+++ device/ble/DeviceBleLightOnoffCctDimHslModeRGB.h
@@ -21,5 +21,5 @@ private:
 	ModuleCct *moduleCct;
 
 public:
-	DeviceBleLightOnoffCctDimHslModeRGB(string id, string name, string mac, Json::Value &dataJson, uint32_t addr, uint32_t type, uint16_t version);
+	DeviceBleLightOnoffCctDimHslModeRGB(string id, string name, string mac, Json::Value &dataJson, uint16_t addr, uint32_t type, uint16_t version);
 };
