+++ device/ble/DeviceBleLightOnoffHslModeRGB.h
@@ -17,5 +17,5 @@ private:
 	ModuleCallScene *moduleCallScene;
 
 public:
-	DeviceBleLightOnoffHslModeRGB(string id, string name, string mac, Json::Value &dataJson, uint32_t addr, uint32_t type, uint16_t version);
+	DeviceBleLightOnoffHslModeRGB(string id, string name, string mac, Json::Value &dataJson, uint16_t addr, uint32_t type, uint16_t version);
 };
