+++ device/ble/DeviceBleRoolDoor.cpp
@@ -1,7 +1,7 @@
 #include "DeviceBleRoolDoor.h"
 #include "Log.h"
 
-DeviceBleRoolDoor::DeviceBleRoolDoor(string id, string name, string mac, Json::Value &dataJson, uint32_t addr, uint32_t type, uint16_t version)
+DeviceBleRoolDoor::DeviceBleRoolDoor(string id, string name, string mac, Json::Value &dataJson, uint16_t addr, uint32_t type, uint16_t version)
 		: DeviceBle(id, name, mac, dataJson, addr, type, version)
 {
 	moduleCurtain = new ModuleCurtain(this, addr);
