+++ device/ble/DeviceBleSensorTempHum.h
@@ -13,5 +13,5 @@ private:
 	ModuleBatteryLevel *moduleBatteryLevel;
 
 public:
-	DeviceBleSensorTempHum(string id, string name, string mac, Json::Value &dataJson, uint32_t addr, uint16_t version);
+	DeviceBleSensorTempHum(string id, string name, string mac, Json::Value &dataJson, uint16_t addr, uint16_t version);
 };
