+++ device/ble/DeviceBleSwitchOnoff.h
@@ -11,5 +11,5 @@ private:
 	ModuleOnOff *moduleOnOff;
 
 public:
-	DeviceBleSwitchOnoff(string id, string name, string mac, Json::Value &dataJson, uint32_t addr, uint32_t type, uint16_t version);
+	DeviceBleSwitchOnoff(string id, string name, string mac, Json::Value &dataJson, uint16_t addr, uint32_t type, uint16_t version);
 };
