+++ device/ble/DeviceBleSwitchScene6DC.cpp
@@ -1,7 +1,7 @@
 #include "DeviceBleSwitchScene6DC.h"
 #include "Log.h"
 
-DeviceBleSwitchScene6DC::DeviceBleSwitchScene6DC(string id, string name, string mac, Json::Value &dataJson, uint32_t addr, uint32_t type, uint16_t version)
+DeviceBleSwitchScene6DC::DeviceBleSwitchScene6DC(string id, string name, string mac, Json::Value &dataJson, uint16_t addr, uint32_t type, uint16_t version)
 		: DeviceBle(id, name, mac, dataJson, addr, type, version)
 {
 	for (int i = 0; i < 6; i++)
