+++ device/ble/DeviceBleSwitchTouch.cpp
@@ -2,7 +2,7 @@
 #include "module/ModuleRelaySwitch.h"
 #include "Log.h"
 
-DeviceBleSwitchTouch::DeviceBleSwitchTouch(string id, string name, string mac, Json::Value &dataJson, uint32_t addr, uint32_t type, uint16_t version, uint8_t numRelay)
+DeviceBleSwitchTouch::DeviceBleSwitchTouch(string id, string name, string mac, Json::Value &dataJson, uint16_t addr, uint32_t type, uint16_t version, uint8_t numRelay)
 		: DeviceBle(id, name, mac, dataJson, addr, type, version)
 {
 	ModuleRelaySwitch *moduleRelaySwitch;
