+++ device/ble/DeviceBleSwitchTouchRgb.cpp
@@ -3,7 +3,7 @@
 #include "module/ModuleRgb.h"
 #include "Log.h"
 
-DeviceBleSwitchTouchRgb::DeviceBleSwitchTouchRgb(string id, string name, string mac, Json::Value &dataJson, uint32_t addr, uint32_t type, uint16_t version, uint8_t countElement)
+DeviceBleSwitchTouchRgb::DeviceBleSwitchTouchRgb(string id, string name, string mac, Json::Value &dataJson, uint16_t addr, uint32_t type, uint16_t version, uint8_t countElement)
 		: DeviceBle(id, name, mac, dataJson, addr, type, version)
 {
 	ModuleOnOff *moduleOnOff;
